update: updates deps and updates core to 8.6.17
This bugfix release is intended to resolve issues that might interfere with future security release upgrades. Symfony 3.4.25 reverted a previous change which affected Drupal's lazy session handling. This release updates Symfony to 3.4.25 and reverts a previous workaround for the regression.

Versions of Drupal 8 prior to 8.6.x are end-of-life and do not receive security coverage. Sites on 8.6.x will receive security coverage until December 4, 2019 (when Drupal 8.8.0 is scheduled for release and 8.6.x security coverage ends).